% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevinQC-package.R
\docType{package}
\name{alevinQC-package}
\alias{alevinQC}
\alias{alevinQC-package}
\title{alevinQC}
\description{
alevinQC
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoneson/alevinQC}
  \item Report bugs at \url{https://github.com/csoneson/alevinQC/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})

Authors:
\itemize{
  \item Avi Srivastava
  \item Rob Patro
  \item Dongze He
}

}
\keyword{internal}
