% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevinQCReport.R
\name{alevinQCReport}
\alias{alevinQCReport}
\title{Generate Alevin summary report}
\usage{
alevinQCReport(baseDir, outputFile, outputDir = "./",
  outputFormat = NULL, showCode = FALSE, forceOverwrite = FALSE,
  sampleId, knitrProgress = FALSE, quiet = FALSE,
  ignorePandoc = FALSE, ...)
}
\arguments{
\item{sampleId}{Sample ID}

\item{basedir}{Base directory (should be the directory containing the
\code{alevin} directory).}
}
\value{
No value is returned, but a report is generated in the
  \code{outputDir} directory.
}
\description{
Generate a report summarizing the main aspects of an Alevin quantification
run. The report generation assumes that Alevin has been run with
--dumpFeatures flag to generate the necessary output files.
}
\details{
When the function is called, an .Rmd template file will be copied
  into the output folder, and \code{rmarkdown::render} will be called to
  generate the final report. If there is already a .Rmd file with the same
  name in the output folder, the function will raise an error and stop, to
  avoid overwriting the existing file. The reason for this behaviour is that
  the copied template in the output folder will be deleted once the report is
  generated.
}
\examples{

}
\author{
Charlotte Soneson
}
