% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevinQCShiny.R
\name{alevinQCShiny}
\alias{alevinQCShiny}
\title{Generate alevin summary shiny app}
\usage{
alevinQCShiny(baseDir, sampleId)
}
\arguments{
\item{baseDir}{Path to the output directory from the alevin run (should be
the directory containing the \code{alevin} directory).}

\item{sampleId}{Sample ID, will be used set the title for the app.}
}
\value{
A shiny app.
}
\description{
Generate a shiny app summarizing the main aspects of an alevin quantification
run. The app generation assumes that alevin has been run with the
--dumpFeatures flag to generate the necessary output files.
}
\examples{
app <- alevinQCShiny(baseDir = system.file("extdata/alevin_example_v0.14",
                                           package = "alevinQC"),
                     sampleId = "example")
if (interactive()) {
    shiny::runApp(app)
}

}
\author{
Charlotte Soneson
}
