% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinHistogram.R
\name{plotAlevinHistogram}
\alias{plotAlevinHistogram}
\title{Histogram of selected summary statistic}
\usage{
plotAlevinHistogram(cbTable, colName = "dedupRate", axisLabel = colName)
}
\arguments{
\item{cbTable}{\code{data.frame} (such as the \code{cbTable} returned by
\code{readAlevinQC}) containing the desired summary statistic in a
column.}

\item{colName}{Character scalar giving the name of a numeric column of
\code{cbTable} to plot.}

\item{axisLabel}{Character scalar giving the label of the selected
statistic (will be displayed as the axis label in the plot).}
}
\value{
A ggplot object
}
\description{
Histogram of selected summary statistic
}
\examples{
alevin <- readAlevinQC(system.file("extdata/alevin_example_v0.14",
                                   package = "alevinQC"))
plotAlevinHistogram(alevin$cbTable, colName = "dedupRate",
                    axisLabel = "Deduplication rate")

}
\author{
Charlotte Soneson
}
