% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinQuant.R
\name{plotAlevinQuant}
\alias{plotAlevinQuant}
\title{Panel of plots with quantification summary statistics}
\usage{
plotAlevinQuant(cbTable)
}
\arguments{
\item{cbTable}{\code{data.frame} (such as the \code{cbTable} returned by
\code{readAlevinQC}) with collapsed barcode frequencies, the total UMI
count and the number of detected genes for each cell.}
}
\value{
A ggplot object
}
\description{
Panel of plots with quantification summary statistics
}
\examples{
alevin <- readAlevinQC(system.file("extdata/alevin/neurons_900",
                                   package = "tximportData"))
plotAlevinQuant(alevin$cbTable)

}
\author{
Charlotte Soneson
}
