% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinQuantPairs.R
\name{plotAlevinQuantPairs}
\alias{plotAlevinQuantPairs}
\title{Plot quantification statistics}
\usage{
plotAlevinQuantPairs(quantbcs)
}
\arguments{
\item{quantbcs}{data.frame with barcode frequencies (as returned by
\code{readAlevinQC}).}
}
\value{
A ggplot object
}
\description{
Plot quantification statistics
}
\author{
Charlotte Soneson
}
