% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlevin.R
\name{readAlevin}
\alias{readAlevin}
\title{Read binary Alevin quant matrix}
\usage{
readAlevin(baseDir = NULL, quiet = TRUE)
}
\arguments{
\item{baseDir}{Path to the output directory from the Alevin run}

\item{quiet}{Logical, whether to suppress printing of the number of detected
molecules.}
}
\value{
A matrix with gene counts for the whitelisted barcodes. Rows
  represent genes, columns cell barcodes.
}
\description{
Read binary Alevin quant matrix
}
\author{
Avi Srivastava
}
