% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlevinFryQC.R
\name{readAlevinFryQC}
\alias{readAlevinFryQC}
\title{Read alevin-fry data required to generate summary report}
\usage{
readAlevinFryQC(mapDir, permitDir, quantDir)
}
\arguments{
\item{mapDir}{Map directory}

\item{permitDir}{Permit list directory}

\item{quantDir}{Quant directory}
}
\value{
A list collecting all necessary information for generating the
    summary report/shiny app.
}
\description{
Read all alevin-fry output files required to generate the
summary report or shiny app.
}
\examples{
#alevin <- readAlevinQC(system.file("extdata/alevin_example_v0.14",
#                       package = "alevinQC"))

}
\author{
Charlotte Soneson
}
