% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevin_report_shiny.R
\name{alevinQCShiny}
\alias{alevinQCShiny}
\title{Generate Alevin summary shiny app}
\usage{
alevinQCShiny(baseDir, sampleId)
}
\arguments{
\item{baseDir}{Path to the output directory from the Alevin run (should be
the directory containing the \code{alevin} directory).}

\item{sampleId}{Sample ID, will be used set the title for the app.}
}
\value{
A shiny app.
}
\description{
Generate a shiny app summarizing the main aspects of an Alevin quantification
run. The app generation assumes that Alevin has been run with the
--dumpFeatures flag to generate the necessary output files.
}
\author{
Charlotte Soneson
}
