% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinBarcodeCollapse.R
\name{plotAlevinBarcodeCollapse}
\alias{plotAlevinBarcodeCollapse}
\title{Plot cell barcode collapsing}
\usage{
plotAlevinBarcodeCollapse(quantbcs)
}
\arguments{
\item{quantbcs}{data.frame with barcode frequencies (as returned by
\code{readAlevinQC}).}
}
\value{
A ggplot object
}
\description{
Plot cell barcode collapsing
}
\author{
Charlotte Soneson
}
