% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinKneeNbrGenes.R
\name{plotAlevinKneeNbrGenes}
\alias{plotAlevinKneeNbrGenes}
\title{Plot nbr genes knee plot}
\usage{
plotAlevinKneeNbrGenes(quantbcs)
}
\arguments{
\item{quantbcs}{data.frame with barcode frequencies (as returned by
\code{readAlevinQC}).}
}
\value{
A ggplot object
}
\description{
Plot nbr genes knee plot
}
\author{
Charlotte Soneson
}
