% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinKneeRaw.R
\name{plotAlevinKneeRaw}
\alias{plotAlevinKneeRaw}
\title{Plot raw knee plot}
\usage{
plotAlevinKneeRaw(rawcbfreq, threshold)
}
\arguments{
\item{rawcbfreq}{data.frame with raw barcode frequencies (as returned by
\code{readAlevinQC}).}

\item{threshold}{The number of initially selected barcodes (will be indicated
in the plot).}
}
\value{
A ggplot object
}
\description{
Plot raw knee plot
}
\author{
Charlotte Soneson
}
