% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinQuantPairs.R
\name{plotAlevinQuantPairs}
\alias{plotAlevinQuantPairs}
\title{Pairs plot with quantification summary statistics}
\usage{
plotAlevinQuantPairs(cbTable)
}
\arguments{
\item{cbTable}{\code{data.frame} (such as the \code{cbTable} returned by
\code{readAlevinQC}) with collapsed barcode frequencies, the total UMI
count and the number of detected genes for each cell.}
}
\value{
A ggmatrix object
}
\description{
Pairs plot with quantification summary statistics
}
\examples{
alevin <- readAlevinQC(system.file("extdata/alevin_example_pre0.14",
                                   package = "alevinQC"))
plotAlevinQuantPairs(alevin$cbTable)

}
\author{
Charlotte Soneson
}
