% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlevinQuantPairs.R
\name{plotAlevinQuantPairs}
\alias{plotAlevinQuantPairs}
\title{Pairs plot with quantification summary statistics}
\usage{
plotAlevinQuantPairs(cbTable, colName = "inFinalWhiteList")
}
\arguments{
\item{cbTable}{\code{data.frame} (such as the \code{cbTable} returned by
\code{readAlevinQC}) with collapsed barcode frequencies, the total UMI
count and the number of detected genes for each cell.}

\item{colName}{Character scalar giving the name of a logical column of
\code{cbTable} to use for coloring the points.}
}
\value{
A ggmatrix object
}
\description{
Pairs plot with quantification summary statistics
}
\examples{
alevin <- readAlevinQC(system.file("extdata/alevin_example_v0.14",
                                   package = "alevinQC"))
plotAlevinQuantPairs(alevin$cbTable, colName = "inFinalWhiteList")

}
\author{
Charlotte Soneson
}
