% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlevinQC.R
\name{readAlevinQC}
\alias{readAlevinQC}
\title{Read all input data}
\usage{
readAlevinQC(baseDir, quiet = TRUE)
}
\arguments{
\item{baseDir}{Path to the output directory from the Alevin run (should be
the directory containing the \code{alevin} directory).}

\item{quiet}{Logical, whether to show progress messages.}
}
\value{
A list with all information necessary to create the alevinQC
  report/app.
}
\description{
Read all input data
}
\author{
Charlotte Soneson
}
