% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlevinQC.R
\name{readAlevinQC}
\alias{readAlevinQC}
\title{Read Alevin data required to generate summary report}
\usage{
readAlevinQC(baseDir)
}
\arguments{
\item{baseDir}{Path to the output directory from the Alevin run (should be
the directory containing the \code{alevin} directory).}
}
\value{
A list collecting all necessary information for generating the
  summary report/shiny app.
}
\description{
Read all Alevin output files required to generate the summary report or shiny
app.
}
\author{
Charlotte Soneson
}
